feat(dashboard): gallery mode
